test: 처음 카테고리 삽입 기능 service에서 config 디렉토리로 이동
